todo_project/requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jwt==1.3.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
psycopg2-binary==2.9.3